fix: lazy-load recharts off first paint (closes launch M1)
